chore(explorer): break apart giant transfer component
